requirements.txt
adafruit-io==2.3.2
certifi==2019.6.16
chardet==3.0.4
idna==2.8
paho-mqtt==1.4.0
pkg-resources==0.0.0
pyserial==3.4
python-firebase==1.2
requests==2.22.0
urllib3==1.25.3
